+++ src/app/domain/enums/index.ts
@@ -1,1 +1,2 @@
 export * from './metric-unit-type';
+export * from './measurement-unit-type';