+++ src/app/domain/enums/measurement-unit-type.ts
@@ -1,0 +1,6 @@
+export enum MeasurementSystem {
+    METRIC = 1,
+    IMPERIAL = 2,
+    SYSTEM = 3,
+    CUSTOM = 4
+  }